correct formation
